Дополнительные функции из модуля main перенесены в модуль utils. Все функции залогированы
