+++ src/reports.py
@@ -1,5 +1,14 @@
-from typing import Any
+import logging
+from typing import Any, Optional
 
+import pandas as pd
+
+logger = logging.getLogger(__name__)
+logger.setLevel(logging.INFO)
+file_handler = logging.FileHandler("../logs/reports.log", "w")
+file_formatter = logging.Formatter("%(asctime)s - %(filename)s - %(levelname)s: %(message)s")
+file_handler.setFormatter(file_formatter)
+logger.addHandler(file_handler)
 
 
 def log(filename: Any = "result.log") -> Any:
@@ -11,6 +20,7 @@ def log(filename: Any = "result.log") -> Any:
         def checking(*args: Any, **kwargs: Any) -> None:
             """Подфункция декоратора, выводит результаты работы функции в файл"""
 
+            logger.info("Используется логирующий декоратор")
             result = func(*args, **kwargs)
             file = open(f"../logs/{filename}", "w")
             file.write(str(result))
@@ -20,17 +30,26 @@ def log(filename: Any = "result.log") -> Any:
 
     return decorator
 
+
 @log()
-def spending_by_category(transactions: list[dict], category: str, date: str = None) -> list[dict]:
+def spending_by_category(main_transactions: pd.DataFrame, category: str, date: Optional[str] = None) -> list[dict]:
+    """Функция выводящая траты по категориям, принимает на вход DataFrame с транзакциями, категорию,
+    и опциональную дату для сортировки. Принимает дату в формате 'DD.MM.YYYY'."""
+
+    logger.info("Запущена функция выводящая траты по категориям и опциональной дате")
     result = []
     output = []
+    transactions = main_transactions.to_dict(orient="records")
     if date is not None:
+        logger.info("Дата введена")
         day = int(date[:2])
         month = int(date[3:5])
         if month <= 3:
+            logger.info("Введён месяц раньше Марта")
             year = int(date[6:10]) - 1
             month = month + 9
         else:
+            logger.info("Введён месяц после Марта")
             year = int(date[6:10])
             month = month - 3
         start_date = f"{day}.0{month}.{year}"
@@ -38,8 +57,10 @@ def spending_by_category(transactions: list[dict], category: str, date: str = No
             if start_date < transaction["Дата операции"] < date:
                 result.append(transaction)
     else:
+        logger.info("Дата не введена")
         result = transactions
     for transaction in result:
         if category == transaction["Категория"]:
             output.append(transaction)
+    logger.info("Завершена функция выводящая данные из Exel")
     return output
